requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==2.0
djangorestframework==3.7.3
idna==2.6
pytz==2017.3
requests==2.18.4
urllib3==1.22
